requirements.txt
# Core web framework and server
flask==3.1.1
gunicorn==23.0.0

# Machine Learning and Data Processing
tensorflow==2.19.0
numpy==2.1.3
Pillow==11.2.1